requirements.txt
numpy
torch
wandb
pytest
torchvision
scikit-learn
torchmetrics
opencv-python
python-dotenv
huggingface_hub
pytorch-lightning